# storage/api-docs/v2-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Brillio API",
        "description": "API Backend for Brillio Orientation System",
        "contact": {
            "email": "contact@brillio.africa"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Brillio API Server"
        }
    ],
    "paths": {
...
        "/api/v2/messages": {
            "get": {
                "tags": [
                    "Messagerie"
                ],
                "summary": "Liste des conversations actives de l'utilisateur",
                "operationId": "1bc87c6531814d7fa948e1d596fc72a2",
                "responses": {
                    "200": {
                        "description": "Liste des conversations"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/messages/{mentorship}": {
            "get": {
                "tags": [
                    "Messagerie"
                ],
                "summary": "Afficher les messages d'une conversation spécifique",
                "operationId": "fcd6ba3c94e9057431d418121069f35a",
                "parameters": [
                    {
                        "name": "mentorship",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des messages de la conversation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messagerie"
                ],
                "summary": "Envoyer un message dans une conversation",
                "operationId": "e510786292ffb144d2a0e4e7ec4d771c",
                "parameters": [
                    {
                        "name": "mentorship",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "Bonjour !"
                                    },
                                    "attachment": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message envoyé avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/messages/file/{message}/download": {
            "get": {
                "tags": [
                    "Messagerie"
                ],
                "summary": "Télécharger la pièce jointe d'un message",
                "operationId": "d7c1425f5888ccd87cdb4b6a999afed3",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier téléchargé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/messages/{message}/update": {
            "patch": {
                "tags": [
                    "Messagerie"
                ],
                "summary": "Modifier le contenu d'un message envoyé",
                "operationId": "16ab03192e075f1013e53bdc71bbe2e0",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body"
                                ],
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "Message modifié"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message mis à jour"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/messages/{message}": {
            "delete": {
                "tags": [
                    "Messagerie"
                ],
                "summary": "Supprimer un message (soft delete)",
                "operationId": "7ebb341d6760f599983d037c154b7ce0",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message supprimé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}